tests/wpt/web-platform-tests/dom/ranges/Range-stringifier.html

<!doctype html>
<meta charset="utf-8">
<title>Range stringifier</title>
<link rel="author" title="KiChjang" href="mailto:kungfukeith11@gmail.com">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id=test>Test div</div>
<div id=log></div>
<script>
test(function() {
  var r = new Range();
  var testDiv = document.getElementById("test");
  r.setStart(testDiv, 0);
  r.setEndAfter(testDiv);
  assert_equals(r.toString(), testDiv.textContent);
});
</script>
